ui(Time Forecast): componentised
