+++ src/lib/components/weather/time/card.svelte
@@ -1,0 +1,14 @@
+<script>
+	import Icon from './icon.svelte';
+	export let currentTime, time;
+</script>
+
+<div
+	class="flex flex-col items-center flex-shrink-0 py-2 space-y-2 text-white rounded-lg w-28 h-36 bg-gradient-to-tr from-black to-gray-800"
+	class:to-primary-600={currentTime === time}
+	class:from-primary-500={currentTime === time}
+>
+	<span class="text-lg font-bold uppercase">{currentTime === time ? 'Now' : time}</span>
+	<Icon />
+	<span class="text-lg font-bold">21°</span>
+</div>
